+++ pokedex1/src/Global/GlobalState.js
