+++ pokedex1/src/components/PokeDetails/PokeMoves.js
