+++ pokedex1/src/components/PokeDetails/PokeType.js
