+++ pokedex1/src/pages/DetailsPage/DetailsPage.js
@@ -1,12 +1,21 @@
-import React from "react";
+import PokeImageDetails from "../../Components/PokeDetails/PokeImage";
+import PokeMoves from "../../Components/PokeDetails/PokeMoves";
+import PokeStats from "../../Components/PokeDetails/PokeStats";
+import { DivMoves, PokeDetailsDiv } from "./StyledDetailsPage";
+import PokeType from "../../Components/PokeDetails/PokeType"
 
 
-function DetailPage(){
-    return(
-        <div>
-            <h1>Detail Page</h1>
-        </div>
+export default function PageDetails () {
+
+    return (
+        <PokeDetailsDiv>
+            <PokeImageDetails/>
+            <PokeStats/>
+            <DivMoves>
+                <PokeType/>
+                <PokeMoves/>
+            </DivMoves>
+        </PokeDetailsDiv>
     )
-}
 
-export default DetailPage;
+}
