+++ pokedex1/src/pages/DetailsPage/StyledDetailsPage.js
