+++ pokedex1/src/pages/HomePage/styledHomePage.js
