+++ pokedex1/src/pages/Route/Router.js
@@ -1,25 +1,29 @@
+import React from "react";
 import { BrowserRouter, Switch, Route } from "react-router-dom";
-import DetailPage from "../DetailsPage/DetailsPage";
-import Home from "../Home/Home";
-import Pokedex from "../Pokedex/Pokedex";
+import Home from "../Pages/HomePage/HomePage";
+import Pokedex from "../Pages/PokedexPage/PokedexPage";
+import PageDetails from "../Pages/DetailsPage/DetailsPage";
+import Header from "../Components/Header/Header";
 
+const Router = () => {
+  return (
+    <BrowserRouter>
+      <Switch>
+        <Route exact path={"/"}>
+          <Header />
+          <Home />
+        </Route>
+        <Route exact path={"/pokedex"}>
+          <Header />
+          <Pokedex />
+        </Route>
+        <Route exact path={"/details"}>
+          <Header />
+          <PageDetails />
+        </Route>
+      </Switch>
+    </BrowserRouter>
+  );
+};
 
-function Router(){
-    return(
-        <BrowserRouter>
-            <Switch>
-                <Route exact path={"/"}>
-                    <Home/>
-                </Route>
-                <Route exact path={"/pokedex"}>
-                    <Pokedex/>
-                </Route>
-                <Route exact path={"/detailpage"}>
-                    <DetailPage/>
-                </Route>
-            </Switch>
-        </BrowserRouter>
-    )
-}
 export default Router;
-
